backend/requirements.txt
flask
gunicorn
mysql-connector-python
pytest
pytest-cov
flake8